@@ -8,12 +8,14 @@
         "northwestern university",
         "laravel"
     ],
-    "authors": [{
-        "name": "Nicholas Evans",
-        "email": "nick.evans@northwestern.edu"
-    }],
+    "authors": [
+        {
+            "name": "Nicholas Evans",
+            "email": "nick.evans@northwestern.edu"
+        }
+    ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "guzzlehttp/guzzle": "^7.0|^6.0",
         "northwestern-sysdev/event-hub-php-sdk": "^3.0",
         "laravel/ui": "^3.0|^2.0|^4.0",
@@ -22,9 +24,9 @@
         "firebase/php-jwt": "^5.3"
     },
     "require-dev": {
-        "orchestra/testbench": "~8.0",
+        "orchestra/testbench": "~9.0",
         "php-coveralls/php-coveralls": "^2.4",
-        "phpunit/phpunit": "^9.0",
+        "phpunit/phpunit": "^10.0",
         "laravel/pint": "^1.13",
         "larastan/larastan": "^2.0"
     },
