@@ -1,14 +1,15 @@
 <?php
 
-namespace Northwestern\SysDev\SOA\Tests\Exceptions;
+namespace Northwestern\SysDev\SOA\Tests\Feature\Exceptions;
 
 use Northwestern\SysDev\SOA\Exceptions\InsecureSsoError;
 use Orchestra\Testbench\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
-class InsecureSsoErrorTest extends TestCase
+final class InsecureSsoErrorTest extends TestCase
 {
-    /** @test */
-    public function throwable()
+    #[Test]
+    public function throwable(): void
     {
         $this->expectExceptionMessageMatches('/https/');
 
