@@ -1,13 +1,13 @@
 <?php
 
-namespace Northwestern\SysDev\SOA\Tests\Auth\Entity;
+namespace Northwestern\SysDev\SOA\Tests\Feature\Auth\Entity;
 
 use Northwestern\SysDev\SOA\Auth\Entity\ActiveDirectoryUser;
 use Orchestra\Testbench\TestCase;
 
-class ActiveDirectoryUserTest extends TestCase
+final class ActiveDirectoryUserTest extends TestCase
 {
-    public function testEntity()
+    public function testEntity(): void
     {
         $user = new ActiveDirectoryUser('abcdefg', [
             'mailNickname' => 'TEST123',
