@@ -1,14 +1,15 @@
 <?php
 
-namespace Northwestern\SysDev\SOA\Tests;
+namespace Northwestern\SysDev\SOA\Tests\Feature;
 
 use Northwestern\SysDev\SOA\DirectorySearch;
+use Northwestern\SysDev\SOA\Tests\TestCase;
 
-class DirectoySearchTest extends TestCase
+final class DirectoySearchTest extends TestCase
 {
     protected $service = DirectorySearch::class;
 
-    public function testGoodLookup()
+    public function testGoodLookup(): void
     {
         $this->api->setHttpClient($this->mockedResponse(200, '{"results":[{ "displayName" : [ "Test E User" ], "givenName" : [ "Test" ], "sn" : [ "User" ], "eduPersonNickname" : [ "test" ], "mail" : "test@example.org", "nuStudentEmail" : "", "title" : [ "Tester" ], "telephoneNumber" : "123 1231234", "nuTelephoneNumber2" : "", "nuTelephoneNumber3" : "", "nuOtherTitle" : "" }]}'));
 
@@ -16,7 +17,7 @@ class DirectoySearchTest extends TestCase
         $this->assertArrayHasKey('mail', $info);
     } // end testGoodLookup
 
-    public function testBadLookup()
+    public function testBadLookup(): void
     {
         $this->api->setHttpClient($this->mockedResponse(404, '{"errorCode":404,"errorMessage":"No Data Found for = uid=test"}'));
 
@@ -24,7 +25,7 @@ class DirectoySearchTest extends TestCase
         $this->assertNotEmpty($this->api->getLastError());
     } // end testBadLookup
 
-    public function testBadPerms()
+    public function testBadPerms(): void
     {
         $this->api->setHttpClient($this->mockedResponse(401, '{"fault":{"faultstring":"Invalid ApiKey for given resource","detail":{"errorcode":"oauth.v2.InvalidApiKeyForGivenResource"}}}'));
 
@@ -32,7 +33,7 @@ class DirectoySearchTest extends TestCase
         $this->assertNotEmpty($this->api->getLastError());
     } // end testBadPerms
 
-    public function testBadApiKey()
+    public function testBadApiKey(): void
     {
         $this->api->setHttpClient($this->mockedResponse(401, '{"fault":{"faultstring":"Failed to resolve API Key variable request.header.apikey","detail":{"errorcode":"steps.oauth.v2.FailedToResolveAPIKey"}}}'));
 
@@ -40,7 +41,7 @@ class DirectoySearchTest extends TestCase
         $this->assertNotEmpty($this->api->getLastError());
     } // end testBadApiKey
 
-    public function testConnectionFailure()
+    public function testConnectionFailure(): void
     {
         $this->api->setHttpClient($this->mockedConnError());
 
