@@ -1,19 +1,19 @@
 <?php
 
-namespace Northwestern\SysDev\SOA\Tests;
+namespace Northwestern\SysDev\SOA\Tests\Feature;
 
 use Northwestern\SysDev\SOA\Providers\NuSoaServiceProvider;
 use Northwestern\SysDev\SOA\Routing\EventHubWebhookRegistration;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 
-class WebhookRouteRegistrationTest extends BaseTestCase
+final class WebhookRouteRegistrationTest extends BaseTestCase
 {
     protected function getPackageProviders($application)
     {
         return [NuSoaServiceProvider::class];
     } // end getPackageProviders
 
-    public function test_route_registration()
+    public function test_route_registration(): void
     {
         $route = app()->router->post('/webhook/foo')->eventHubWebhook('foo.my-queue');
 
@@ -21,7 +21,7 @@ class WebhookRouteRegistrationTest extends BaseTestCase
         $this->assertEquals(1, count($registered_hooks));
     } // end test_route_registration
 
-    public function test_uses_hmac_when_configured()
+    public function test_uses_hmac_when_configured(): void
     {
         $secret = 'abcdefg';
         $this->app['config']->set('nusoa.eventHub.hmacVerificationSharedSecret', $secret);
@@ -37,7 +37,7 @@ class WebhookRouteRegistrationTest extends BaseTestCase
         $this->assertEquals($secret, $hook['webhookSecurity'][0]['secretKey']);
     } // end test_uses_hmac_when_configured
 
-    public function test_custom_security_setup()
+    public function test_custom_security_setup(): void
     {
         $this->app['config']->set('nusoa.eventHub.hmacVerificationSharedSecret', null);
 
@@ -53,7 +53,7 @@ class WebhookRouteRegistrationTest extends BaseTestCase
         $this->assertEquals($secret, $hook['webhookSecurity'][0]['apiKey']);
     } // end test_custom_security_setup
 
-    public function test_multiple_security_modes()
+    public function test_multiple_security_modes(): void
     {
         $this->app['config']->set('nusoa.eventHub.hmacVerificationSharedSecret', 'hmac-key');
 
@@ -67,7 +67,7 @@ class WebhookRouteRegistrationTest extends BaseTestCase
         $this->assertEquals(2, count($hook['webhookSecurity']));
     } // end test_multiple_security_modes
 
-    public function test_change_content_type()
+    public function test_change_content_type(): void
     {
         $content_type = 'application/xml';
         $route = app()->router->post('/webhook/foo')->eventHubWebhook('foo.my-queue', ['contentType' => $content_type]);
